Remove export duplication for Flags
