+++ src/actions/resize/ResizeSimpleAction.ts
@@ -2,7 +2,7 @@ import {Qualifier} from "../../internal/qualifier/Qualifier";
 import {Action} from "../../internal/Action";
 import {toFloatAsString} from "../../internal/utils/toFloatAsString";
 import {AspectRatioQualifierValue} from "../../values/aspectRatio/AspectRatioQualifierValue";
-import {regionRelative, relative} from "../../values/flag/Flag";
+import {regionRelative, relative} from "../../values/flag";
 import {FlagQualifier} from "../../values/flag/FlagQualifier";
 
 /**
