--- src/values/flag/custom.ts
@@ -1,12 +1,0 @@
-import {FlagQualifier} from './FlagQualifier';
-
-/**
- * @memberOf Values.Flag
- * @description Allows custom flag
- * @return {Values.Flag.FlagQualifier}
- */
-function custom(value:string): FlagQualifier {
-  return new FlagQualifier(value);
-}
-
-export default custom;
