--- src/values/flag/progressive.ts
@@ -1,23 +1,0 @@
-import {FlagQualifier} from './FlagQualifier';
-
-/**
- * @memberOf Values.Flag
- * @description Generates a JPG image using the progressive (interlaced) JPG format.
- *
- * This format allows the browser to quickly show a low-quality rendering of the image until the full-quality
- * image is loaded.
- *
- * @param {string} mode The mode to determine a specific progressive outcome as follows:
- * * semi - A smart optimization of the decoding time, compression level and progressive rendering
- *          (less iterations). This is the default mode when using q_auto.
- * * steep - Delivers a preview very quickly, and in a single later phase improves the image to
- *           the required resolution.
- * * none  - Use this to deliver a non-progressive image. This is the default mode when setting
- *           a specific value for quality.
- * @return {Values.Flag.FlagQualifier}
- */
-function progressive(mode?: string): FlagQualifier {
-  return new FlagQualifier('progressive', mode);
-}
-
-export {progressive};
